refactor: restructured entire project into modular feature folders
- Group components by domains into dashboard/, insights/ and landing/
- Implemented dashboard container component and enable routing between landing page and dashboard.
- Removed empty App.css
- Replaced three duplicated formatDuration implementations and scattered date/colour constants with shared lib/format and lib/theme
- Extracted logic into lib/: format helpers, theme constants and a dedicated analytics module per feature (heatmap, marathons, obsessions and discovery)
