+++ client/src/App.tsx
@@ -1,15 +1,7 @@
 import { useSpotifyAuth } from './hooks/useSpotifyAuth';
 import { buildSpotifyAuthUrl } from './api/spotify';
-import { LandingPage } from './components/LandingPage';
-import { TopTrackCard } from './components/TopTrackCard';
-import { TrackList } from './components/TrackList';
-import { PopularityBarChart } from './components/PopularityBarChart';
-import { RecentPlayCount } from './components/RecentPlayCount';
-import { GenrePieChart } from './components/GenrePieChart';
-import { ListeningHeatmap } from './components/ListeningHeatmap';
-import { ListeningMarathons } from './components/ListeningMarathons';
-import { ArtistObsessionPhases } from './components/ArtistObsessionPhases';
-import { DiscoveryRateChart } from './components/DiscoveryRateChart';
+import { LandingPage } from './components/landing/LandingPage';
+import { Dashboard } from './components/dashboard/Dashboard';
 
 export default function App() {
   const {
@@ -40,35 +32,13 @@ export default function App() {
       {!isLoggedIn ? (
         <LandingPage loginUrl={buildSpotifyAuthUrl()} />
       ) : (
-        <div className="mx-auto max-w-6xl space-y-8">
-          <div className="grid gap-6 md:grid-cols-3">
-            {topTracks[0] && <TopTrackCard track={topTracks[0]} />}
-            <div className="relative md:col-span-1">
-              <div className="hide-scrollbar absolute inset-0 overflow-y-auto">
-                <TrackList tracks={topTracks.slice(1, 20)} />
-              </div>
-            </div>
-            <div className="space-y-6">
-              <PopularityBarChart tracks={topTracks.slice(0, 20)} />
-              <GenrePieChart genres={genreCounts} />
-            </div>
-          </div>
-
-          <div>
-            <h2 className="mb-4 text-sm font-medium uppercase tracking-widest text-zinc-400">
-              Recent Play Count
-            </h2>
-            <RecentPlayCount tracks={topTracks} playCounts={playCounts} />
-          </div>
-
-          <ListeningHeatmap plays={recentPlays} />
-
-          <ListeningMarathons plays={recentPlays} />
-
-          <ArtistObsessionPhases plays={recentPlays} topArtists={topArtists} />
-
-          <DiscoveryRateChart plays={recentPlays} />
-        </div>
+        <Dashboard
+          topTracks={topTracks}
+          topArtists={topArtists}
+          recentPlays={recentPlays}
+          playCounts={playCounts}
+          genreCounts={genreCounts}
+        />
       )}
     </div>
   );
