+++ client/src/components/dashboard/PopularityBarChart.tsx
@@ -1,5 +1,11 @@
 import { BarChart, Bar, XAxis, YAxis, Tooltip, ResponsiveContainer, Cell } from 'recharts';
-import type { SpotifyTrack } from '../types/spotify';
+import type { SpotifyTrack } from '../../types/spotify';
+import {
+  SPOTIFY_GREEN,
+  NEUTRAL_FILL,
+  TOOLTIP_CONTENT_STYLE,
+  TOOLTIP_LABEL_STYLE,
+} from '../../lib/theme';
 
 export function PopularityBarChart({ tracks }: { tracks: SpotifyTrack[] }) {
   const data = tracks.map((track) => ({
@@ -17,14 +23,14 @@ export function PopularityBarChart({ tracks }: { tracks: SpotifyTrack[] }) {
           <XAxis dataKey="name" hide />
           <YAxis domain={[0, 100]} tick={{ fill: '#71717a', fontSize: 11 }} />
           <Tooltip
-            contentStyle={{ background: '#18181b', border: 'none', borderRadius: 8 }}
-            labelStyle={{ color: '#fff', fontSize: 12 }}
-            itemStyle={{ color: '#1db954' }}
+            contentStyle={TOOLTIP_CONTENT_STYLE}
+            labelStyle={TOOLTIP_LABEL_STYLE}
+            itemStyle={{ color: SPOTIFY_GREEN }}
             formatter={(value: number) => [`${value}/100`, 'Popularity']}
           />
           <Bar dataKey="popularity" radius={[4, 4, 0, 0]}>
             {data.map((_, i) => (
-              <Cell key={i} fill={i === 0 ? '#1db954' : '#3f3f46'} />
+              <Cell key={i} fill={i === 0 ? SPOTIFY_GREEN : NEUTRAL_FILL} />
             ))}
           </Bar>
         </BarChart>
