+++ client/src/components/dashboard/TopTrackCard.tsx
@@ -1,10 +1,5 @@
-import type { SpotifyTrack } from '../types/spotify';
-
-function formatDuration(ms: number): string {
-  const minutes = Math.floor(ms / 60000);
-  const seconds = Math.floor((ms % 60000) / 1000);
-  return `${minutes}:${seconds.toString().padStart(2, '0')}`;
-}
+import type { SpotifyTrack } from '../../types/spotify';
+import { formatTrackLength } from '../../lib/format';
 
 export function TopTrackCard({ track }: { track: SpotifyTrack }) {
   return (
@@ -26,7 +21,9 @@ export function TopTrackCard({ track }: { track: SpotifyTrack }) {
       <p className="mt-0.5 text-sm text-zinc-400">
         {track.artists.map((a) => a.name).join(', ')}
       </p>
-      <p className="mt-2 text-xs text-zinc-500">Duration: {formatDuration(track.duration_ms)}</p>
+      <p className="mt-2 text-xs text-zinc-500">
+        Duration: {formatTrackLength(track.duration_ms)}
+      </p>
     </a>
   );
 }
