+++ client/src/components/dashboard/TrackList.tsx
@@ -1,10 +1,5 @@
-import type { SpotifyTrack } from '../types/spotify';
-
-function formatDuration(ms: number): string {
-  const minutes = Math.floor(ms / 60000);
-  const seconds = Math.floor((ms % 60000) / 1000);
-  return `${minutes}:${seconds.toString().padStart(2, '0')}`;
-}
+import type { SpotifyTrack } from '../../types/spotify';
+import { formatTrackLength } from '../../lib/format';
 
 export function TrackList({ tracks }: { tracks: SpotifyTrack[] }) {
   return (
@@ -29,7 +24,9 @@ export function TrackList({ tracks }: { tracks: SpotifyTrack[] }) {
               {track.artists.map((a) => a.name).join(', ')}
             </p>
           </div>
-          <span className="shrink-0 text-xs text-zinc-500">{formatDuration(track.duration_ms)}</span>
+          <span className="shrink-0 text-xs text-zinc-500">
+            {formatTrackLength(track.duration_ms)}
+          </span>
         </a>
       ))}
     </div>
