+++ client/src/components/insights/DiscoveryRateChart.tsx
@@ -9,110 +9,16 @@ import {
   CartesianGrid,
   ResponsiveContainer,
 } from 'recharts';
-import type { RecentPlay } from '../types/spotify';
+import type { RecentPlay } from '../../types/spotify';
+import { analyseDiscovery } from '../../lib/discovery';
+import {
+  SPOTIFY_GREEN,
+  NEUTRAL_FILL,
+  TOOLTIP_CONTENT_STYLE,
+  TOOLTIP_LABEL_STYLE,
+} from '../../lib/theme';
 
-const DAY_MS = 24 * 60 * 60 * 1000;
-// Above this history span the chart switches from daily to weekly buckets.
-const WEEKLY_SPAN_DAYS = 21;
-
-interface Bucket {
-  key: number;
-  label: string;
-  new: number;
-  replayed: number;
-  rate: number;
-}
-
-function startOfDay(ts: number): number {
-  const date = new Date(ts);
-  date.setHours(0, 0, 0, 0);
-  return date.getTime();
-}
-
-function startOfWeek(ts: number): number {
-  const date = new Date(startOfDay(ts));
-  const mondayOffset = (date.getDay() + 6) % 7;
-  date.setDate(date.getDate() - mondayOffset);
-  return date.getTime();
-}
-
-function formatLabel(ts: number): string {
-  return new Date(ts).toLocaleDateString(undefined, {
-    month: 'short',
-    day: 'numeric',
-  });
-}
-
-interface DiscoverySummary {
-  buckets: Bucket[];
-  weekly: boolean;
-  overallRate: number;
-  newTracks: number;
-  newArtists: number;
-  totalPlays: number;
-}
-
-function analyseDiscovery(plays: RecentPlay[]): DiscoverySummary {
-  const sorted = plays
-    .filter((p) => !Number.isNaN(new Date(p.played_at).getTime()))
-    .sort(
-      (a, b) =>
-        new Date(a.played_at).getTime() - new Date(b.played_at).getTime(),
-    );
-
-  const empty: DiscoverySummary = {
-    buckets: [],
-    weekly: false,
-    overallRate: 0,
-    newTracks: 0,
-    newArtists: 0,
-    totalPlays: 0,
-  };
-  if (sorted.length === 0) return empty;
-
-  const first = new Date(sorted[0].played_at).getTime();
-  const last = new Date(sorted[sorted.length - 1].played_at).getTime();
-  const weekly = (last - first) / DAY_MS > WEEKLY_SPAN_DAYS;
-
-  const seenTracks = new Set<string>();
-  const seenArtists = new Set<string>();
-  const buckets = new Map<number, Bucket>();
-
-  for (const play of sorted) {
-    const ts = new Date(play.played_at).getTime();
-    const key = weekly ? startOfWeek(ts) : startOfDay(ts);
-    const bucket =
-      buckets.get(key) ??
-      ({ key, label: formatLabel(key), new: 0, replayed: 0, rate: 0 } as Bucket);
-
-    const trackId = play.track?.id;
-    if (trackId && !seenTracks.has(trackId)) {
-      seenTracks.add(trackId);
-      bucket.new += 1;
-    } else {
-      bucket.replayed += 1;
-    }
-    for (const artist of play.track?.artists ?? []) {
-      if (artist.id) seenArtists.add(artist.id);
-    }
-    buckets.set(key, bucket);
-  }
-
-  const ordered = [...buckets.values()].sort((a, b) => a.key - b.key);
-  for (const bucket of ordered) {
-    const total = bucket.new + bucket.replayed;
-    bucket.rate = total === 0 ? 0 : Math.round((bucket.new / total) * 100);
-  }
-
-  return {
-    buckets: ordered,
-    weekly,
-    overallRate: Math.round((seenTracks.size / sorted.length) * 100),
-    newTracks: seenTracks.size,
-    newArtists: seenArtists.size,
-    totalPlays: sorted.length,
-  };
-}
+const DISCOVERY_LINE = '#8b5cf6';
 
 export function DiscoveryRateChart({ plays }: { plays: RecentPlay[] }) {
   const { buckets, weekly, overallRate, newTracks, newArtists, totalPlays } =
@@ -158,8 +64,8 @@ export function DiscoveryRateChart({ plays }: { plays: RecentPlay[] }) {
             tickFormatter={(value: number) => `${value}%`}
           />
           <Tooltip
-            contentStyle={{ background: '#18181b', border: 'none', borderRadius: 8 }}
-            labelStyle={{ color: '#fff', fontSize: 12 }}
+            contentStyle={TOOLTIP_CONTENT_STYLE}
+            labelStyle={TOOLTIP_LABEL_STYLE}
             itemStyle={{ fontSize: 12 }}
             cursor={{ fill: 'rgba(255,255,255,0.04)' }}
             formatter={(value: number, name: string) =>
@@ -176,23 +82,23 @@ export function DiscoveryRateChart({ plays }: { plays: RecentPlay[] }) {
             dataKey="new"
             name="New"
             stackId="plays"
-            fill="#1db954"
+            fill={SPOTIFY_GREEN}
           />
           <Bar
             yAxisId="count"
             dataKey="replayed"
             name="Replayed"
             stackId="plays"
-            fill="#3f3f46"
+            fill={NEUTRAL_FILL}
             radius={[4, 4, 0, 0]}
           />
           <Line
             yAxisId="rate"
             dataKey="rate"
             name="Discovery rate"
-            stroke="#8b5cf6"
+            stroke={DISCOVERY_LINE}
             strokeWidth={2}
-            dot={{ r: 3, fill: '#8b5cf6' }}
+            dot={{ r: 3, fill: DISCOVERY_LINE }}
           />
         </ComposedChart>
       </ResponsiveContainer>
