+++ client/src/lib/format.ts
@@ -1,0 +1,48 @@
+/** Formatting helpers shared across the app. */
+
+/** Track length as m:ss, e.g. 187000 -> "3:07". */
+export function formatTrackLength(ms: number): string {
+  const minutes = Math.floor(ms / 60000);
+  const seconds = Math.floor((ms % 60000) / 1000);
+  return `${minutes}:${seconds.toString().padStart(2, '0')}`;
+}
+
+/** Longer duration as "Xh Ym", or "Ym" when under an hour. */
+export function formatDuration(ms: number): string {
+  const totalMinutes = Math.round(ms / 60000);
+  const hours = Math.floor(totalMinutes / 60);
+  const minutes = totalMinutes % 60;
+  return hours === 0 ? `${minutes}m` : `${hours}h ${minutes}m`;
+}
+
+/** Hour of day as a 12-hour label, e.g. 15 -> "3pm". */
+export function formatHour(hour: number): string {
+  if (hour === 0) return '12am';
+  if (hour === 12) return '12pm';
+  return hour < 12 ? `${hour}am` : `${hour - 12}pm`;
+}
+
+/** Short calendar date, e.g. "May 12". */
+export function formatShortDate(ts: number): string {
+  return new Date(ts).toLocaleDateString(undefined, {
+    month: 'short',
+    day: 'numeric',
+  });
+}
+
+/** Weekday and date, e.g. "Mon, May 12". */
+export function formatWeekdayDate(ts: number): string {
+  return new Date(ts).toLocaleDateString(undefined, {
+    weekday: 'short',
+    month: 'short',
+    day: 'numeric',
+  });
+}
+
+/** Clock time, e.g. "3:45 PM". */
+export function formatClockTime(ts: number): string {
+  return new Date(ts).toLocaleTimeString(undefined, {
+    hour: 'numeric',
+    minute: '2-digit',
+  });
+}
